Fixes from class (sonar, id long, abstract entity, create update date, simplified service methods, changed post to put)

+++ backend-projekt-web/src/main/java/com/backendprojektweb/repository/HallRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HallRepository extends JpaRepository<Hall, Integer> {
+public interface HallRepository extends JpaRepository<Hall, Long> {
 }
