+++ backend-projekt-web/src/main/java/com/backendprojektweb/repository/ReservationSeatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservationSeatRepository extends JpaRepository<ReservationSeat, Integer> {
+public interface ReservationSeatRepository extends JpaRepository<ReservationSeat, Long> {
 }
